==> R/s9_compute_coef_bootstrap.R <==
# s9_compute_coef_bootstrap.R


# Compute regression coefficient bootstrap to calculate the standard error.


# Project : habitat_suitability_curves_fda
# Author  : Jeremie Boudreault
# Email   : Jeremie.Boudreault [at] ete.inrs [dot] ca
# Depends : R (v3.6.3)
# License : CC BY-NC-ND 4.0



# Library ----------------------------------------------------------------------


library(data.table)
library(ggplot2)
library(FDboost)


# Functions --------------------------------------------------------------------


source(file.path("R", "functions", "internals.R"))
source(file.path("R", "functions", "fdboost_helpers.R"))


# Parameters -------------------------------------------------------------------


# Number of bootstrap.
n_bs <- 1000L


# Imports ----------------------------------------------------------------------


# Fitted functional models.
frm_results <- qs::qread(
    file = file.path("out", "tmp", "s7_frm_results_full.qs")
)

# Functional observations.
fd_curves_list <- qs::qread(
    file = file.path("out", "tmp", "s5_fd_curves_list.qs")
)


# Generate bootstrap sample ----------------------------------------------------


generate_bs_sample <- function(l) {
    n_obs <- nrow(l$Y)
    boot_i <- sample(seq_len(n_obs), size = n_obs, replace = TRUE)
    l$Y <- l$Y[boot_i, ]
    l$X <- l$X[boot_i, ]
    return(l)
}


# Compute bootstrap of coefficients --------------------------------------------


# Apply over all variables.
bs_res <- lapply(
    X = names(var_names),
    FUN = function(var) {

        # Apply over all bootstrap (n_bs).
        do.call(what = abind, args = lapply(
            X   = seq_len(n_bs),
            FUN = function(i) {

                # Fit functional model.
                obj <- .FDboost(
                    data         = generate_bs_sample(fd_curves_list[[var]]),
                    fdboost_opts = frm_results[[var]]$fdboost_opts,
                    mstop_best   = frm_results[[var]]$mstop_best
                )

                # Fitted coefficients.
                coef_fit <- coef(obj$fit)

                # Extract matrix.
                coef_mat <- coef_fit$smterms$`bsignal(X) %O% bbs(s)`$value

                # Add row and column names.
                rownames(coef_mat) <- coef_fit$smterms$`bsignal(X) %O% bbs(s)`$y
                colnames(coef_mat) <- coef_fit$smterms$`bsignal(X) %O% bbs(s)`$x

                # Return coef_mat.
                return(coef_mat)

            }
        ))
    }
)

# Add names to the results.
names(bs_res) <- names(var_names)


# Standard error of the coefficients -------------------------------------------


# Calculate standard error.
std <- lapply(bs_res, function(w) apply(w, FUN = sd, MARGIN = c(1L, 2L)))

# Melt into a data.table.
std_dt <- data.table::setDT(reshape2::melt(std))

# Update names.
names(std_dt) <- c("Y", "X", "Z", "MODEL")

# Replace values of <MODEL> with variables names and units.
std_dt[, MODEL := factor(
    x      = unlist(var_names_u[std_dt$MODEL], use.names = FALSE),
    levels = unlist(var_names_u)
)]


# Plot standard error of each regression coefficient ---------------------------


.plot_std <- function(data) {

    p <- ggplot(
        data = data,
        mapping = aes(
            x = X,
            y = Y,
            z = Z,
            fill = Z
        )
    ) +
    geom_raster(
        alpha = 1L
    ) +
    geom_abline(
        slope     = 1L,
        intercept = 0L,
        lty       = 2L,
        lwd       = 0.1
    ) +
    scale_fill_gradient(
        low     = "#FFFFFF00",
        high    = "darkgrey",
        guide   = guide_colourbar(barwidth = 12L)
    ) +
    scale_x_continuous(
        expand = c(0.001, 0.001)
    ) +
    scale_y_continuous(
        expand = c(0.001, 0.001)
    ) +
    labs(
        y    = "",
        x    = "",
        fill = ""
    ) +
    facet_wrap(
        facets = ~ MODEL,
        nrow   = 1L,
        ncol   = 1L,
        scales = "free",
    ) +
    theme(
        panel.border     = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.line        = element_line(colour = "white")
    ) +
    legend_bottom
    return(p)

}

# Generate all plots.
ps <- lapply(var_names_u, function(var) .plot_std(std_dt[MODEL == var, ]))

# Add x and y labs.
ps[[1L]] <- ps[[1L]] + ylab(hab_names[["SELECTED"]])
ps[[2L]] <- ps[[2L]] + xlab(hab_names[["AVAILABLE"]])


# Plot results -----------------------------------------------------------------


# Save as a pdf for future use.
pdf(
    file   = file.path("out", "plots", "fig_7_frm_coef_bootstrap.pdf"),
    width  = 10L,
    height = 4L
)

# Plot.
ggpubr::ggarrange(
    plotlist = ps,
    ncol     = 3L
)

# Save plot.
dev.off()


# Exports ----------------------------------------------------------------------


qs::qsave(
    x    = bs_res,
    file = file.path("out", "tmp", "s9_frm_coef_bootstrap.qs")
)
